@@ -14,78 +15,79 @@
   },
   "scripts": {
     "dev": "vite dev",
-    "dev:s": "vite dev --host --https",
     "build": "npm run webkit && npm run get-schema && vite build",
+    "get-schema": "node scripts/schema.js --run",
     "preview": "vite preview",
     "test": "npm run test:integration && npm run test:unit",
     "check": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json",
     "check:watch": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json --watch",
-    "lint": "prettier --plugin-search-dir . --check . && eslint .",
-    "format": "prettier --plugin-search-dir . --write .",
+    "lint": "prettier --check . && eslint .",
+    "format": "prettier --write .",
     "test:integration": "playwright test",
     "test:unit": "vitest",
-    "storybook": "storybook dev -p 4000 --ci",
+    "storybook": "npm run webkit && storybook dev -p 4000 --ci --disable-telemetry",
     "build-storybook": "storybook build",
-    "webkit": "node scripts/webkit",
-    "prepare": "husky install && npm run webkit",
-    "patch:svelte-kit": "node scripts/patch",
-    "get-schema": "node scripts/schema.js --run"
-  },
-  "dependencies": {
-    "@popsql/monaco-sql-languages": "^0.0.3",
-    "@sveltejs/adapter-node": "^1.3.1",
-    "@vitejs/plugin-basic-ssl": "^1.1.0",
-    "chart.js": "^4.3.3",
-    "esm-env": "^1.0.0",
-    "marked": "^6.0.0",
-    "medium-editor": "^5.23.3",
-    "monaco-editor": "^0.41.0",
-    "turndown": "^7.1.2",
-    "ua-parser-js": "^1.0.35"
+    "webkit": "node scripts/webkit"
   },
   "devDependencies": {
-    "@playwright/test": "^1.35.1",
-    "@storybook/addon-essentials": "^7.3.2",
-    "@storybook/addon-interactions": "^7.3.2",
-    "@storybook/addon-links": "^7.3.2",
-    "@storybook/blocks": "^7.3.2",
-    "@storybook/svelte": "^7.3.2",
-    "@storybook/sveltekit": "^7.3.2",
-    "@storybook/testing-library": "^0.2.0",
-    "@sveltejs/kit": "^1.25.1",
-    "@types/ua-parser-js": "^0.7.36",
-    "@typescript-eslint/eslint-plugin": "^6.4.0",
-    "@typescript-eslint/parser": "^6.4.0",
-    "eslint": "^8.46.0",
-    "eslint-config-prettier": "^9.0.0",
-    "eslint-plugin-storybook": "^0.6.13",
-    "eslint-plugin-svelte": "^2.32.4",
-    "esm-env": "^1.0.0",
-    "graphql": "^16.7.1",
-    "husky": "^8.0.3",
-    "lint-staged": "^13.2.3",
-    "miragejs": "^0.1.47",
-    "node-fetch": "^3.3.2",
-    "prettier": "^2.8.8",
-    "prettier-plugin-svelte": "^2.10.1",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
-    "san-webkit": "github:santiment/san-webkit#f98cf6c",
-    "sass": "^1.65.1",
-    "storybook": "^7.3.2",
-    "svelte": "^4.2.8",
-    "svelte-check": "^3.5.0",
-    "svelte-preprocess": "^5.0.4",
+    "@chromatic-com/storybook": "^1.2.18",
+    "@melt-ui/pp": "^0.3.0",
+    "@melt-ui/svelte": "^0.76.0",
+    "@playwright/test": "^1.42.0",
+    "@storybook/addon-essentials": "^8.0.6",
+    "@storybook/addon-interactions": "^8.0.6",
+    "@storybook/addon-links": "^8.0.6",
+    "@storybook/addon-themes": "^8.0.6",
+    "@storybook/blocks": "^8.0.6",
+    "@storybook/svelte": "^8.0.6",
+    "@storybook/sveltekit": "^8.0.6",
+    "@storybook/test": "^8.0.6",
+    "@sveltejs/adapter-auto": "^3.2.0",
+    "@sveltejs/kit": "^2.5.5",
+    "@sveltejs/vite-plugin-svelte": "^3.1.0",
+    "@types/eslint": "^8.56.5",
+    "@types/ua-parser-js": "^0.7.39",
+    "@typescript-eslint/eslint-plugin": "^7.1.0",
+    "@typescript-eslint/parser": "^7.1.0",
+    "autoprefixer": "^10.4.18",
+    "eslint": "^9.0.0",
+    "eslint-config-prettier": "^9.1.0",
+    "eslint-plugin-storybook": "^0.8.0",
+    "eslint-plugin-svelte": "2.36.0-next.13",
+    "estree-walker": "^3.0.3",
+    "magic-string": "^0.30.8",
+    "postcss": "^8.4.35",
+    "prettier": "^3.2.5",
+    "prettier-plugin-svelte": "^3.2.2",
+    "sass": "^1.72.0",
+    "storybook": "^8.0.6",
+    "svelte": "5.0.0-next.99",
+    "svelte-check": "^3.6.9",
     "svelte-preprocess-cssmodules": "github:DmitriVanGuard/svelte-preprocess-cssmodules#3b24f22",
-    "tslib": "^2.6.0",
-    "typescript": "^5.1.6",
-    "vite": "^4.4.9",
-    "vitest": "^0.33.0"
+    "tailwindcss": "^3.4.1",
+    "tslib": "^2.6.2",
+    "typescript": "^5.4.5",
+    "vite": "^5.2.8",
+    "vitest": "^1.3.1"
   },
-  "type": "module",
+  "dependencies": {
+    "@popsql/monaco-sql-languages": "^0.1.0",
+    "@sveltejs/adapter-node": "^5.0.1",
+    "@xyflow/svelte": "^0.0.39",
+    "chart.js": "^4.4.2",
+    "esm-env": "^1.0.0",
+    "monaco-editor": "^0.47.0",
+    "prettier-plugin-tailwindcss": "^0.5.11",
+    "rxjs": "^7.8.1",
+    "san-webkit": "github:santiment/san-webkit#15afa8a",
+    "svelte-runes": "^0.0.3",
+    "tailwind-merge": "^2.2.1",
+    "ua-parser-js": "^1.0.37",
+    "zod": "^3.22.4"
+  },
   "pnpm": {
     "patchedDependencies": {
-      "@sveltejs/kit@1.25.1": "patches/@sveltejs__kit@1.25.1.patch"
+      "@sveltejs/kit@2.5.5": "patches/@sveltejs__kit@2.5.5.patch"
     }
   }
 }
