@@ -1,29 +1,35 @@
 <script lang="ts">
-  import { queryGenerateTitleBySql } from '$lib/EntityHead/Query/api'
-  import { noop } from 'san-webkit/lib/utils'
+  import { useObserve } from 'svelte-runes'
+  import { tap } from 'rxjs'
+  import { useAiSuggestion } from '../flow/ai'
 
-  export let sql: string
-  export let label = 'Use AI'
-  export let onSuggestion = noop as (
-    suggestion: Awaited<ReturnType<typeof queryGenerateTitleBySql>>,
-  ) => void
+  let {
+    sql,
+    label,
+    onSuggestion,
+  }: { sql: string; label?: string; onSuggestion: (data: any) => void } = $props()
 
-  let loading = false
+  let loading = $state(false)
+
+  const { data$, getAiTitleAndDescription } = useAiSuggestion()
+
+  useObserve(() =>
+    data$.pipe(
+      tap(() => (loading = false)),
+      tap(onSuggestion),
+    ),
+  )
 
   function onClick() {
     loading = true
-    queryGenerateTitleBySql(sql)
-      .then(onSuggestion)
-      .finally(() => {
-        loading = false
-      })
+    getAiTitleAndDescription(sql)
   }
 </script>
 
 {#if loading}
   <div class="loading-spin" />
 {:else}
-  <button class="expl-tooltip" aria-label={label} on:click={onClick}> ✨ </button>
+  <button class="expl-tooltip" aria-label={label} onclick={onClick}> ✨ </button>
 {/if}
 
 <style>
