@@ -1,11 +1,12 @@
+/** @type { import("eslint").Linter.Config } */
 module.exports = {
   root: true,
   extends: [
     'eslint:recommended',
-    // 'plugin:@typescript-eslint/recommended',
+    'plugin:@typescript-eslint/recommended',
     'plugin:svelte/recommended',
     'prettier',
-    // 'plugin:storybook/recommended',
+    'plugin:storybook/recommended',
   ],
   parser: '@typescript-eslint/parser',
   plugins: ['@typescript-eslint'],
@@ -19,37 +20,17 @@ module.exports = {
     es2017: true,
     node: true,
   },
-  globals: {
-    SAN: true,
-    App: true,
-  },
   overrides: [
     {
       files: ['*.svelte'],
       parser: 'svelte-eslint-parser',
-      globals: {
-        SAN: true,
-        App: true,
-      },
       parserOptions: {
         parser: '@typescript-eslint/parser',
-      },
-
-      rules: {
-        'css-unused-selector': 'off',
-        'svelte/valid-compile': 'off',
       },
     },
   ],
 
   rules: {
-    'css-unused-selector': 'off',
-    'no-unused-vars': 'off',
-    '@typescript-eslint/no-unused-vars': [
-      'error',
-      {
-        argsIgnorePattern: '^_',
-      },
-    ],
+    '@typescript-eslint/no-namespace': 'off',
   },
 }
