@@ -1,8 +1,11 @@
 <script>
-  import LoginPrompt from 'webkit/ui/LoginPrompt/index.svelte'
+  import LoginPrompt from 'san-webkit/lib/ui/LoginPrompt/index.svelte'
   import { getAuthCtx } from '../ctx'
+  import { page } from '$app/stores'
 
   const { onMetamaskClick } = getAuthCtx()
+
+  const successRedirect = $page.url.searchParams.get('from') ?? ''
 </script>
 
 <LoginPrompt
@@ -11,5 +14,6 @@
   bottomLabel="Have an account?"
   bottomAction="Log in"
   bottomHref="/login"
+  from={successRedirect}
   {onMetamaskClick}
 />
