@@ -1,38 +1,22 @@
-import type { PageLoad } from '../../$types'
-
-import { BROWSER } from 'esm-env'
 import { redirect } from '@sveltejs/kit'
-import { getIdFromSEOLink } from 'webkit/utils/url'
-import { queryGetDashboard, queryGetLegacyDashboard } from '$lib/api/dashboard/get'
+import { getIdFromSEOLink } from 'san-webkit/lib/utils/url'
+import { UniQuery } from '$lib/api/index.js'
+import { queryGetDashboard } from './api.js'
 
 export const ssr = false
 
-export const load: PageLoad = async (event) => {
-  const { slug } = event.params
+export const load = async (event) => {
+  const { slug = 'new' } = event.params
 
   if (slug === 'new') return
 
-  const id = getIdFromSEOLink(slug)
-  if (Number.isInteger(id) === false) {
+  const dashboardId = getIdFromSEOLink(slug)
+
+  if (Number.isInteger(dashboardId) === false) {
     throw redirect(302, '/dashboard/new')
   }
 
-  const apiDashboard = await queryGetDashboard(id, event as App.RequestEvent)
-
-  if (BROWSER) {
-    const isLegacy =
-      apiDashboard.panels?.length &&
-      !(apiDashboard.queries.length || apiDashboard.textWidgets.length)
-
-    if (isLegacy) {
-      apiDashboard.panels = (await queryGetLegacyDashboard(id)).panels
-      apiDashboard.isLegacy = true
-    }
-  }
-
-  if (!apiDashboard.description) {
-    apiDashboard.description = ''
-  }
+  const apiDashboard = await queryGetDashboard(UniQuery(event.fetch))(dashboardId)
 
   return {
     apiDashboard,
