@@ -1,27 +1,29 @@
 <script lang="ts">
   import type { LayoutData } from './$types'
 
+  import '../app.css'
+
   import { BROWSER } from 'esm-env'
-  import { Customer$$ } from 'webkit/stores/customer'
-  import { CurrentUser$$ } from 'webkit/stores/user'
-  import { Device$$ } from 'webkit/stores/responsive'
-  import { newAppTooltipsCtx } from 'webkit/ui/Tooltip/ctx'
-  import { UI$$ } from 'webkit/stores/ui'
-  import Dialogs from 'webkit/ui/Dialog/Dialogs.svelte'
-  import Notifications from 'webkit/ui/Notifications'
-  import FeatureWalkthrough from 'webkit/ui/FeatureWalkthrough/svelte'
-  import OnlyOnDevice from 'webkit/ui/OnlyOnDevice.svelte'
-  import NavHeader from '$lib/NavHeader/index.svelte'
-  import Tracking from './Tracking.svelte'
+  import { Customer$$ } from 'san-webkit/lib/stores/customer'
+  import { CurrentUser$$ } from 'san-webkit/lib/stores/user'
+  import { Device$$ } from 'san-webkit/lib/stores/responsive'
+  import { newAppTooltipsCtx } from 'san-webkit/lib/ui/Tooltip/ctx'
+  import { UI$$ } from 'san-webkit/lib/stores/ui'
+  import Dialogs from 'san-webkit/lib/ui/Dialog/Dialogs.svelte'
+  import Notifications from 'san-webkit/lib/ui/Notifications'
+  import FeatureWalkthrough from 'san-webkit/lib/ui/FeatureWalkthrough/svelte'
+  import OnlyOnDevice from 'san-webkit/lib/ui/OnlyOnDevice.svelte'
+  import NavHeader from '$lib/ui/NavHeader'
   import Gdpr from '$lib/Gdpr.svelte'
+  // import Tracking from './Tracking.svelte'
 
   export let data: LayoutData
 
+  newAppTooltipsCtx()
   CurrentUser$$(data.session.currentUser)
   Customer$$(data.session.customer)
   UI$$()
   const { device$ } = Device$$(data.session.device)
-  newAppTooltipsCtx()
 </script>
 
 <svelte:head>
@@ -34,21 +36,19 @@
   <NavHeader />
 </OnlyOnDevice>
 
-<screen class="row">
+{#if BROWSER}
   <Gdpr>
     <slot />
   </Gdpr>
-</screen>
 
-{#if BROWSER}
   <Dialogs />
   <FeatureWalkthrough />
   <Notifications />
 
-  <Tracking />
+  <!-- <Tracking /> -->
 {/if}
 
-<style lang="scss">
+<style lang="postcss">
   :root {
     --accent: var(--green);
     --accent-hover: var(--green-hover);
@@ -58,19 +58,13 @@
     scroll-behavior: smooth;
   }
 
-  /* TODO: Is it needed? [@vanguard | 10 Jan, 2023] */
   :global(body) {
-    position: relative;
-    display: flex;
-    flex-direction: column;
-    min-height: 100vh;
+    @apply flex min-h-[100vh] flex-col;
   }
 
-  screen {
-    flex: 1;
-  }
-
-  @include dac(tablet, phone, phone-xs) {
-    padding-bottom: 83px;
-  }
+  /*
+   @include dac(tablet, phone, phone-xs) {
+     padding-bottom: 83px;
+   }
+   */
 </style>
