@@ -1,6 +1,6 @@
 <script lang="ts">
-  import { getCustomer$Ctx } from 'webkit/stores/customer'
-  import { getCurrentUser$Ctx } from 'webkit/stores/user'
+  import { getCustomer$Ctx } from 'san-webkit/lib/stores/customer'
+  import { getCurrentUser$Ctx } from 'san-webkit/lib/stores/user'
   // import { LoginType } from 'webkit/analytics/events/general'
   import { goto } from '$app/navigation'
   import { mutateVerifyEmail } from './api'
@@ -14,14 +14,14 @@
   const { customer$ } = getCustomer$Ctx()
 
   if (process.browser) {
-    const { email, token } = data
+    const { email, token, successRedirect } = data
 
     mutateVerifyEmail(email as string, token as string)
       .then((currentUser) => {
         currentUser$.set(currentUser)
         customer$.refetch()
 
-        goto('/')
+        goto(successRedirect)
 
         // trackSignupLogin(currentUser.firstLogin, LoginType.EMAIL)
       })
