@@ -1,17 +1,25 @@
+import { preprocessMeltUI, sequence } from '@melt-ui/pp'
 import adapter from '@sveltejs/adapter-node'
-import webkitConfig from 'san-webkit/svelte.config.js'
+import { vitePreprocess } from '@sveltejs/vite-plugin-svelte'
+import { componentStyleSelector } from 'san-webkit/svelte.config.js'
 
-/** @type {import('@sveltejs/kit').Config} */
+/** @type {import('@sveltejs/kit').Config}*/
 const config = {
-  ...webkitConfig,
-
+  // Consult https://kit.svelte.dev/docs/integrations#preprocessors
+  // for more information about preprocessors
+  preprocess: sequence([vitePreprocess(), preprocessMeltUI(), componentStyleSelector()]),
   kit: {
+    // adapter-auto only supports some environments, see https://kit.svelte.dev/docs/adapter-auto for a list.
+    // If your environment is not supported or you settled on a specific environment, switch out the adapter.
+    // See https://kit.svelte.dev/docs/adapters for more information about adapters.
     adapter: adapter(),
     alias: {
       '$routes/*': './src/routes/*',
       '$static/*': './static/*',
     },
   },
+  compilerOptions: {
+    hmr: false,
+  },
 }
-
 export default config
